requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-genai
python-dotenv
beautifulsoup4
requests
markdown
